requirements.txt
numpy
torch
torchvision
opencv-python
tqdm
Pillow
rembg
git+https://github.com/facebookresearch/segment-anything.git
